Move file mock-data.js into tests folder
